# static/swagger.yaml
openapi: 3.0.0
info:
  title: Route Optimizer API
  description: |
    API service for optimizing delivery routes with AI/ML techniques.
    This API provides endpoints for geocoding, route optimization, traffic prediction,
    fuel consumption estimation, and other logistics-related services.
  version: 1.0.0
  contact:
    name: API Support
    email: test1@test.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: Main API server
...
paths:
...
  /geocode:
    post:
      summary: Geocode addresses
      description: Convert addresses to geographic coordinates (latitude and longitude)
      tags:
        - geocoding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                addresses:
                  type: array
                  items:
                    type: string
                  example: ["123 Main St, New York, NY", "456 Park Ave, Miami, FL"]
              required:
                - addresses
      responses:
        '200':
          description: Successful geocoding response
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: object
                    additionalProperties:
                      oneOf:
                        - type: object
                          properties:
                            latitude:
                              type: number
                              format: float
                              example: 40.7128
                            longitude:
                              type: number
                              format: float
                              example: -74.0060
                        - type: "null"
                  count:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 2
                      successful:
                        type: integer
                        example: 2
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /distance-matrix:
    post:
      summary: Calculate distance matrix
      description: Calculate distances between multiple coordinates
      tags:
        - geocoding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                coordinates:
                  type: array
                  items:
                    type: object
                    properties:
                      lat:
                        type: number
                        format: float
                        example: 40.7128
                      lng:
                        type: number
                        format: float
                        example: -74.0060
                    required:
                      - lat
                      - lng
              required:
                - coordinates
      responses:
        '200':
          description: Distance matrix
          content:
            application/json:
              schema:
                type: object
                properties:
                  matrix:
                    type: array
                    items:
                      type: array
                      items:
                        type: number
                        format: float
                    example: [[0, 2.5, 3.2], [2.5, 0, 1.8], [3.2, 1.8, 0]]
                  size:
                    type: integer
                    example: 3
                  unit:
                    type: string
                    example: kilometers
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /optimize-route:
    post:
      summary: Optimize delivery routes
      description: Generate optimized routes for deliveries using AI/ML techniques
      tags:
        - optimization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                depot:
                  $ref: '#/components/schemas/Location'
                locations:
                  type: array
                  items:
                    $ref: '#/components/schemas/Location'
                vehicle_count:
                  type: integer
                  example: 1
                  description: Number of vehicles to use
                max_distance:
                  type: number
                  example: 100
                  description: Maximum distance per vehicle (in kilometers)
                use_traffic:
                  type: boolean
                  default: false
                  description: Whether to consider real-time traffic data
                use_fuel_efficient:
                  type: boolean
                  default: false
                  description: Whether to optimize for fuel efficiency
                optimization_objective:
                  type: string
                  enum: [distance, time, fuel, balanced]
                  default: balanced
                  description: The main objective to optimize for
              required:
                - depot
                - locations
      responses:
        '200':
          description: Optimized routes
          content:
            application/json:
              schema:
                type: object
                properties:
                  routes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Route'
                  optimization_type:
                    type: string
                    enum: [distance, traffic_aware, fuel_efficient]
                    example: fuel_efficient
                  vehicle_count:
                    type: integer
                    example: 1
                  total_locations:
                    type: integer
                    example: 10
                  timestamp:
                    type: string
                    format: date-time
                  traffic_info:
                    type: object
                    description: Only present when use_traffic is true
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /predict-travel-time:
    post:
      summary: Predict travel times
      description: Predict travel times based on distances, optionally considering time of day
      tags:
        - optimization
        - traffic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                distances:
                  oneOf:
                    - type: array
                      items:
                        type: number
                        format: float
                      example: [10.5, 15.2, 8.7]
                    - type: array
                      items:
                        type: array
                        items:
                          type: number
                          format: float
                time_of_day:
                  oneOf:
                    - type: string
                      example: "08:30"
                    - type: array
                      items:
                        type: string
                day_of_week:
                  oneOf:
                    - type: string
                      enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
                    - type: integer
                      minimum: 0
                      maximum: 6
                    - type: array
                      items:
                        type: string
              required:
                - distances
      responses:
        '200':
          description: Predicted travel times
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      times:
                        type: array
                        items:
                          type: number
                          format: float
                        example: [0.35, 0.51, 0.29]
                      unit:
                        type: string
                        example: hours
                  - type: object
                    properties:
                      time_matrix:
                        type: array
                        items:
                          type: array
                          items:
                            type: number
                            format: float
                      unit:
                        type: string
                        example: hours
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /traffic-data:
    get:
      summary: Get traffic data
      description: Get traffic data for a geographic area
      tags:
        - traffic
      parameters:
        - name: min_lat
          in: query
          required: true
          schema:
            type: number
            format: float
          example: 40.70
        - name: min_lon
          in: query
          required: true
          schema:
            type: number
            format: float
          example: -74.10
        - name: max_lat
          in: query
          required: true
          schema:
            type: number
            format: float
          example: 40.80
        - name: max_lon
          in: query
          required: true
          schema:
            type: number
            format: float
          example: -73.90
      responses:
        '200':
          description: Traffic data for the specified area
          content:
            application/json:
              schema:
                type: object
                properties:
                  traffic_data:
                    type: object
                    properties:
                      traffic_signals:
                        type: array
                        items:
                          type: object
                      road_speeds:
                        type: object
                      congestion_areas:
                        type: array
                        items:
                          type: object
                      is_simulated:
                        type: boolean
                  bounds:
                    type: array
                    items:
                      type: number
                    example: [40.70, -74.10, 40.80, -73.90]
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Location:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Warehouse"
        address:
          type: string
          example: "123 Main St, New York, NY"
        latitude:
          type: number
          format: float
          example: 40.7128
        longitude:
          type: number
          format: float
          example: -74.0060
        time_window_start:
          type: string
          example: "08:00"
          description: Start of delivery time window (HH:MM format)
        time_window_end:
          type: string
          example: "18:00"
          description: End of delivery time window (HH:MM format)
      required:
        - latitude
        - longitude
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid input parameters"
      required:
        - error